pyproject.toml
[project]
name = "project-file-platform"
version = "0.1.0"
description = "Project-isolated file platform with FastAPI + FastMCP"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]

[project.scripts]
project-file-platform = "project_file_platform:main"
pfp-api = "project_file_platform.api.main:run"
pfp-mcp = "project_file_platform.mcp.server:run"
pfp-migrate = "project_file_platform.api.migrations:run_cli"

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
